requirements.txt
requests
tqdm
pandas
numpy
cvxpy
pyyaml
prettytable
scipy